+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/navigation-next
 
+## 0.0.6
+- [patch] Add debug prop to NavigationProvider that enables Nav API debug logging [018d77d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/018d77d)
+
 ## 0.0.5
 - [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
   - @atlaskit/icon@11.3.0